+++ artemis/src/main/java/com/artemis/systems/IntervalEntitySystem.java
@@ -33,14 +33,16 @@ public abstract class IntervalEntitySystem extends EntitySystem {
 		this.interval = interval;
 	}
 
-
 	@Override
 	protected boolean checkProcessing() {
+		if (intervalDelta > 0 && acc == 0)
+			intervalDelta = 0;
+
 		acc += getTimeDelta();
 		if(acc >= interval) {
-			acc -= interval;
-			intervalDelta = (acc - intervalDelta);
-			
+			intervalDelta = acc;
+			acc = 0;
+
 			return true;
 		}
 		return false;
@@ -52,7 +54,7 @@ public abstract class IntervalEntitySystem extends EntitySystem {
 	 * @return Time passed since last process round.
 	 */
 	protected float getIntervalDelta() {
-		return interval + intervalDelta;
+		return intervalDelta + acc;
 	}
 	
 	protected float getTimeDelta() {